requirements.txt
discord.py
requests
django-apscheduler
python-dotenv
django4-background-tasks
Pybooru